Day 17/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
assemblyai
aiofiles
google-generativeai
pydantic
SpeechRecognition
gTTS
websockets  